refactor - user-api auth controller 병합
