+++ .idea/compiler.xml
@@ -8,10 +8,10 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.36/5a30490a6e14977d97d9c73c924c1f1b5311ea95/lombok-1.18.36.jar" />
         </processorPath>
-        <module name="es3-backend.user-api.main" />
-        <module name="es3-backend.order-api.main" />
+        <module name="com.es3.es3-backend.order-api.main" />
+        <module name="com.es3.es3-backend.user-api.main" />
+        <module name="com.es3.es3-backend.gateway.main" />
         <module name="Backend copy.es3-backend.main" />
-        <module name="es3-backend.gateway.main" />
       </profile>
     </annotationProcessing>
   </component>
