+++ es3-backend/order-api/src/main/java/com/es3/order/OrderApiApplication.java
@@ -2,8 +2,10 @@ package com.es3.order;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class OrderApiApplication {
     public static void main(String[] args) {
         SpringApplication.run(OrderApiApplication.class, args);
