+++ es3-backend/user-api/src/main/java/com/es3/user/common/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.es3.user.auth.dto.common.response;
+package com.es3.user.common.dto.response;
 
 public record TokenResponse(
         String accessToken,
