--- es3-backend/user-api/src/main/java/com/es3/user/config/exception/BannerException.java
@@ -1,7 +1,0 @@
-package com.es3.user.config.exception;
-
-public class BannerException extends AbstractException{
-    public BannerException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
