--- es3-backend/user-api/src/main/java/com/es3/user/config/exception/OrderException.java
@@ -1,7 +1,0 @@
-package com.es3.user.config.exception;
-
-public class OrderException extends AbstractException {
-    public OrderException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
