--- es3-backend/user-api/src/main/java/com/es3/user/config/exception/PaymentException.java
@@ -1,7 +1,0 @@
-package com.es3.user.config.exception;
-
-public class PaymentException extends AbstractException{
-    public PaymentException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
