--- es3-backend/user-api/src/main/java/com/es3/user/config/exception/StoreException.java
@@ -1,7 +1,0 @@
-package com.es3.user.config.exception;
-
-public class StoreException extends AbstractException {
-    public StoreException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
