+++ es3-backend/user-api/src/main/java/com/es3/user/security/EncryptionUtil.java
@@ -1,4 +1,4 @@
-package com.es3.user.auth.security;
+package com.es3.user.security;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
