+++ es3-backend/user-api/src/main/java/com/es3/user/security/JwtFilter.java
@@ -1,4 +1,4 @@
-package com.es3.user.auth.security;
+package com.es3.user.security;
 
 import com.sun.security.auth.UserPrincipal;
 import jakarta.servlet.FilterChain;
