+++ es3-backend/user-api/src/main/java/com/es3/user/seller/dto/request/SellerSignInForm.java
@@ -1,4 +1,4 @@
-package com.es3.user.auth.dto.seller.request;
+package com.es3.user.seller.dto.request;
 
 public record SellerSignInForm(
 	String email,
