+++ es3-backend/user-api/src/main/java/com/es3/user/seller/service/SellerUpdateService.java
@@ -1,11 +1,11 @@
 package com.es3.user.seller.service;
 
-import com.es3.user.auth.dto.seller.request.SellerUpdateRequest;
 import com.es3.user.config.exception.AuthException;
 import com.es3.user.config.exception.ErrorCode;
 import com.es3.user.seller.domain.Seller;
 import com.es3.user.seller.domain.SellerRepository;
 import com.es3.user.seller.dto.SellerDto;
+import com.es3.user.seller.dto.request.SellerUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
