+++ es3-backend/user-api/src/main/java/com/es3/user/user/dto/request/UserSignUpForm.java
@@ -1,4 +1,4 @@
-package com.es3.user.auth.dto.user.request;
+package com.es3.user.user.dto.request;
 
 public record UserSignUpForm(
         String email,
